+++ lib/src/parser/models/flavors/windows.dart
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2024 Angelo Cassano
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import 'package:flutter_flavorizr/src/parser/models/flavors/commons/dummy_assets_mixin.dart';
+import 'package:flutter_flavorizr/src/parser/models/flavors/commons/icon_mixin.dart';
+import 'package:flutter_flavorizr/src/parser/models/flavors/commons/os.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+part 'windows.g.dart';
+
+@JsonSerializable(anyMap: true, createToJson: false)
+class Windows extends OS with IconMixin, DummyAssetsMixin {
+  Windows({
+    bool generateDummyAssets = true,
+    String? icon,
+  }) {
+    this.generateDummyAssets = generateDummyAssets;
+    this.icon = icon;
+  }
+
+  factory Windows.fromJson(Map<String, dynamic> json) =>
+      _$WindowsFromJson(json);
+}